+++ src/Client.php
@@ -16,7 +16,8 @@ namespace Akamai\Open\EdgeGrid;
 use Akamai\Open\EdgeGrid\Handler\Authentication as AuthenticationHandler;
 use Akamai\Open\EdgeGrid\Handler\Debug as DebugHandler;
 use Akamai\Open\EdgeGrid\Handler\Verbose as VerboseHandler;
-use Psr\Log\LoggerInterface;
+use GuzzleHttp\Client as GuzzleClient;
+use GuzzleHttp\ClientTrait;
 
 /**
  * Akamai {OPEN} EdgeGrid Client for PHP
@@ -32,9 +33,11 @@ use Psr\Log\LoggerInterface;
  *
  * @package Akamai\Open\EdgeGrid\Client
  */
-class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
+class Client implements \Psr\Log\LoggerAwareInterface, \Psr\Http\Client\ClientInterface
 {
-    public const VERSION = '2.0.0';
+    use ClientTrait;
+
+    public const VERSION = '2.1.0';
 
     /**
      * @const int Default Timeout in seconds
@@ -54,6 +57,11 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
      * @var bool|resource Whether debug mode is enabled
      */
     protected static $staticDebug = false;
+
+    /**
+     * @var \GuzzleHttp\Client
+     */
+    protected $guzzler;    
 
     /**
      * @var \Akamai\Open\EdgeGrid\Authentication
@@ -95,9 +103,9 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
     protected $debugOverride = false;
 
     /**
-     * @var LoggerInterface|null
+     * @var \Psr\Log\LoggerInterface|null
      */
-    protected ?LoggerInterface $logger = null;
+    protected ?\Psr\Log\LoggerInterface $logger = null;
 
     /**
      * @var string
@@ -119,7 +127,41 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
         $config['headers']['User-Agent'] = 'Akamai-Open-Edgegrid-PHP/' .
             self::VERSION . ' ' . \GuzzleHttp\default_user_agent();
 
-        parent::__construct($config);
+        $this->guzzler = new GuzzleClient($config);
+    }
+
+    /**
+     * Get client configuration options.
+     *
+     * @param string|null $option The config option to retrieve or all if null
+     * 
+     * @return mixed
+     */
+    public function getConfig($option = null)
+    {
+        return $this->guzzler->getConfig($option);
+    }
+
+    /**
+     * Make a request
+     *
+     * @param string $method
+     * @param string $uri
+     * @param array $options
+     * @return \Psr\Http\Message\ResponseInterface
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function request(string $method, $uri = null, array $options = []): \Psr\Http\Message\ResponseInterface
+    {
+        $options = $this->setRequestOptions($options);
+
+        $query = parse_url($uri, PHP_URL_QUERY);
+        if (!empty($query)) {
+            $uri = substr($uri, 0, (strlen($query) + 1) * -1);
+            parse_str($query, $options['query']);
+        }
+        
+        return $this->guzzler->request($method, $uri, $options);
     }
 
     /**
@@ -141,7 +183,34 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
             parse_str($query, $options['query']);
         }
 
-        return parent::requestAsync($method, $uri, $options);
+        return $this->guzzler->requestAsync($method, $uri, $options);
+    }
+
+    /**
+     * Send an HTTP request
+     *
+     * @param \Psr\Http\Message\RequestInterface $request The HTTP request
+     * @param array $options Request options
+     *
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function send(\Psr\Http\Message\RequestInterface $request, array $options = []): \Psr\Http\Message\ResponseInterface
+    {
+        $options = $this->setRequestOptions($options);
+
+        return $this->guzzler->send($request, $options);
+    }
+
+    /**
+     * Send an HTTP request
+     *
+     * @param \Psr\Http\Message\RequestInterface $request The HTTP request
+     *
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function sendRequest(\Psr\Http\Message\RequestInterface $request): \Psr\Http\Message\ResponseInterface
+    {
+        return $this->guzzleClient->sendRequest($request);
     }
 
     /**
@@ -156,7 +225,21 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
     {
         $options = $this->setRequestOptions($options);
 
-        return parent::sendAsync($request, $options);
+        return $this->guzzler->sendAsync($request, $options);
+    }
+
+    /**
+     * Create and send an HTTP OPTIONS request.
+     * Implemented here as it is no longer provided by Guzzle built-in.
+     *
+     * @param string|UriInterface $uri     URI object or string.
+     * @param array               $options Request options to apply.
+     *
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function options($uri, array $options = []): \Psr\Http\Message\ResponseInterface
+    {
+        return $this->request('OPTIONS', $uri, $options);
     }
 
     /**
@@ -269,11 +352,21 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
     /**
      * Set a PSR-3 compatible logger (or use monolog by default)
      *
-     * @param LoggerInterface $logger
-     * @return void
+     * @param \Psr\Log\LoggerInterface $logger
      */
-    public function setLogger(LoggerInterface $logger): void
-    {
+    public function setLogger(
+        \Psr\Log\LoggerInterface $logger = null,
+        $messageFormat = null
+    ): void  {
+        if ($logger === null) {
+            $handler = new \Monolog\Handler\ErrorLogHandler(\Monolog\Handler\ErrorLogHandler::SAPI);
+            $handler->setFormatter(new \Monolog\Formatter\LineFormatter('%message%'));
+            $logger = new \Monolog\Logger('HTTP Log', [$handler]);
+        }
+
+        if ($messageFormat !== null) {
+            $this->setMessageFormat($messageFormat);
+        }
         $formatter = new \GuzzleHttp\MessageFormatter($this->messageFormat);
 
         $handler = \GuzzleHttp\Middleware::log($logger, $formatter);
@@ -281,13 +374,6 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
 
         $handlerStack = $this->getConfig('handler');
         $this->setLogHandler($handlerStack, $handler);
-    }
-
-    public function defaultLogger(): LoggerInterface
-    {
-        $handler = new \Monolog\Handler\ErrorLogHandler(\Monolog\Handler\ErrorLogHandler::SAPI);
-        $handler->setFormatter(new \Monolog\Formatter\LineFormatter('%message%'));
-        return new \Monolog\Logger('HTTP Log', [$handler]);
     }
 
     /**
@@ -307,8 +393,7 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
         $handler->setFormatter(new \Monolog\Formatter\LineFormatter('%message%'));
         $log = new \Monolog\Logger('HTTP Log', [$handler]);
 
-        $this->setMessageFormat($format);
-        $this->setLogger($log);
+        $this->setLogger($log, $format);
 
         return $this;
     }
@@ -529,7 +614,7 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
             $this->config[$what] = $value;
         };
 
-        $closure = $closure->bindTo($this, \GuzzleHttp\Client::class);
+        $closure = $closure->bindTo($this->guzzler, \GuzzleHttp\Client::class);
         $closure();
     }
 
@@ -678,7 +763,6 @@ class Client extends \GuzzleHttp\Client implements \Psr\Log\LoggerAwareInterface
 
             $formatter = new \GuzzleHttp\MessageFormatter($this->messageFormat);
             $handler = \GuzzleHttp\Middleware::log($this->logger, $formatter);
-
 
             $this->setLogHandler($options['handler'], $handler);
             return $options;